% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slSim.R
\name{make_SL_fleet}
\alias{make_SL_fleet}
\title{Make a Short-Lived Species Fleet Object}
\usage{
make_SL_fleet(
  Name = "A fleet",
  nYear = 10,
  pYear = 10,
  Seasons = 12,
  CurrentYear = 2026,
  nSim = 4,
  rec_age = 4,
  nages = 24,
  Effort = NA,
  sel50 = 6,
  selSLP = 2
)
}
\arguments{
\item{Name}{Character string. The name of the stock object. E.g., "North Pacific Neon Flying Squid"}

\item{nYear}{Positive Integer. The number of historical years. E.g., 10, (2015 - 2024)}

\item{pYear}{Positive Integer. The number of projection years. E.g., 10 (2025 - 2034)}

\item{Seasons}{Positive Integer. The number of sub-year time steps. E.g., 12 for a monthly model}

\item{CurrentYear}{Integer. The Calendar year of the last historical year. E.g. 2024.}

\item{nSim}{Positive Integer. The number of independent simulations. E.g., 4 for testing, 24 for preliminary results, 48 for representative results, 192+ for informing management.}

\item{rec_age}{Positive Integer. The age the fish is recruited (in seasonal time steps). E.g., given a monthly model (Seasons = 12) a value of 4 means that the stock recruitment function calculates recruiment into the time step 4 months later.}

\item{nages}{Positive Integer. The number of seasonal time steps that population dynamics will be calculated for. E.g., 24 would be two years in a monthly model. Note that this only necessary if creatures live that long or dynamics are not suitably approximated with a plus group.}

\item{Effort}{A matrix of positive real numbers nSim x time steps (nYear x Seasons). Given a default q value of 1 this is the apical fishing mortality rate. Defaul is 'NA' and in this case effort pattern is simulated with a seasonal and temporal trend.}

\item{sel50}{Positive real number. The age (in seasons) that 50% of individuals are selected, in a logistic selectivity model. E.g., 6 in a monthly model would be 50% selected after 6 months.}

\item{selSLP}{Positive real number. The slope of the logistic selectivity model. E.g., 2 in a monthly model.}

\item{Species}{Character string. The genus and species. E.g., "Ommastrephes bartramii"}

\item{CommonName}{Character string. The common name. E.g., "Neon Flying Squid"}
}
\value{
An object of MSEtool class fleet
}
\description{
A function that builds the fleet component of an operating model for a short-lived species. Requires from
specifications for growth, stock-recruitment, natural mortality rate and spatial distribution. This
is a wrapper function and produces a fully customizable MSEtool object of class 'stock'. You can edit
that object to include very complex stock dynamics if desired.
}
\examples{
A_short_lived_fleet<- make_SL_fleet("Default fleet values for demo")
class(A_short_lived_fleet)
}
\seealso{
\link{make_SL_stock} for making a short-lived stock object and \link{spec_SL_om} for specifying the entire operating model from fleet and stock objects.
}
\author{
T. Carruthers
}
