% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slSim.R
\name{make_SL_stock}
\alias{make_SL_stock}
\title{Make a Short-Lived Species Stock Object}
\usage{
make_SL_stock(
  Name = "A short-lived creature",
  Species = "Shortus liveus",
  CommonName = "Short-lived creature",
  nYear = 10,
  pYear = 10,
  Seasons = 12,
  CurrentYear = 2026,
  nSim = 4,
  rec_age = 4,
  nages = 24,
  PlusGroup = F,
  spawndist = c(0, 0, 0.1, 0.5, 0.3, 0.2, 0, 0, 0, 0, 0, 0),
  Linf = 1,
  K = 0.2,
  t0 = 0,
  Len_CV = 0.2,
  a = 1e-05,
  b = 3,
  M = 0.2,
  amat50 = 6,
  amatSLP = 2,
  h = 0.9,
  sigmaR = 1,
  trunc_sigmaR = 2,
  R_AC = 0.5,
  R0 = 1e+06,
  nareas = 2,
  Frac_area = matrix(c(0.95, 0.9, 0.85, 0.8, 0.75, 0.7, 0.65, 0.6, 0.55, 0.5, 0.5, 0.5,
    0.45, 0.4, 0.35, 0.3, 0.25, 0.2, 0.15, 0.1, 0.05), nrow = 1),
  prob_stay = 0.9
)
}
\arguments{
\item{Name}{Character string. The name of the stock object. E.g., "North Pacific Neon Flying Squid"}

\item{Species}{Character string. The genus and species. E.g., "Ommastrephes bartramii"}

\item{CommonName}{Character string. The common name. E.g., "Neon Flying Squid"}

\item{nYear}{Positive Integer. The number of historical years. E.g., 10, (2015 - 2024)}

\item{pYear}{Positive Integer. The number of projection years. E.g., 10 (2025 - 2034)}

\item{Seasons}{Positive Integer. The number of sub-year time steps. E.g., 12 for a monthly model}

\item{CurrentYear}{Integer. The Calendar year of the last historical year. E.g. 2024.}

\item{nSim}{Positive Integer. The number of independent simulations. E.g., 4 for testing, 24 for preliminary results, 48 for representative results, 192+ for informing management.}

\item{rec_age}{Positive Integer. The age the fish is recruited (in seasonal time steps). E.g., given a monthly model (Seasons = 12) a value of 4 means that the stock recruitment function calculates recruiment into the time step 4 months later.}

\item{nages}{Positive Integer. The number of seasonal time steps that population dynamics will be calculated for. E.g., 24 would be two years in a monthly model. Note that this only necessary if creatures live that long or dynamics are not suitably approximated with a plus group.}

\item{PlusGroup}{Boolean. Should population numbers be aggregated in the nages age class or just be assumed to die off at nages+1?}

\item{spawndist}{Numeric vector Seasons long. The seasonal distribution of spawning. E.g., c(0, 0, 0.1, 0.5, 0.3, 0.2, 0, 0, 0, 0, 0, 0), is a monthly spawning pattern where 50 per cent of spawning occurs in April.}

\item{Linf}{Positive real number or vector of positive real numbers nSim long. The asymptotic length.}

\item{K}{Positive real number or vector of positive real numbers nSim long. The somatic growth rate (von B.) per season.}

\item{t0}{Negative real number or vector of negative real numbers nSim long. The theoretical length at age (seasonal) zero.}

\item{Len_CV}{Positive real number. The coefficient of variation of length at age. Typically between 0.05 and 0.25.}

\item{a}{Positive real number. Weight (W) at length (L) parameter a. W = aL^b. Converts units of length to weight.}

\item{b}{Positive real number. Weight (W) at length (L) parameter b. W = aL^b. Typically approximately cubic (3-ish)}

\item{M}{Positive real number or vector of positive real numbers nSim long. The instantaneous rate of natural mortality per season.}

\item{amat50}{Positive real number or vector of positive real numbers nSim long. The age (in seasons) that 50% of individuals are mature, in a logistic maturity model. E.g., 6 in a monthly model would be 50% mature after 6 months.}

\item{amatSLP}{Positive real number or vector of positive real numbers nSim long. The slope of the logistic maturity model. E.g., 2 in a monthly model.}

\item{h}{Positive real number or vector of positive real numbers nSim long. Steepness of the stock-recruitment relationship (Bev-Holt)}

\item{sigmaR}{Positive real number or vector of positive real numbers nSim long. The lognormal standard deviation in recruitment deviations.}

\item{trunc_sigmaR}{Positive real number. The number of standard deviations to truncate recruitment deviations. E.g., 2 resamples recruitment deviations if over two standard deviations from the mean (upper tail q of 1.96).}

\item{R_AC}{Positive real number or vector of positive real numbers nSim long. The lag-1 autocorrelation in recruitment deviations.}

\item{R0}{Positive real number or vector of positive real numbers nSim long. Unfished recruitment (number of individuals).}

\item{nareas}{Positive real number. The number of spatial areas. Should match the spec of Frac_area which has nareas-1 rows.}

\item{Frac_area}{Matrix of fractions nareas-1 x nages. The fraction of stock numbers found in each area by age. Frac_area[1,3] = 0.9 means that 90 percent of individuals of age 3 are found in area 1.}

\item{prob_stay}{Fraction. The tendency for individuals to remain in the same area among seasonal time steps. Using gravity equations, the function optimizes for movement that obtains Frac_area while staying as close to prob_stay as possible.}
}
\value{
An object of MSEtool class stock
}
\description{
A function that builds the stock operating model component for a short-lived species from
specifications for growth, stock-recruitment, natural mortality rate and spatial distribution. This
is a wrapper function and produces a fully customizable MSEtool object of class 'stock'. You can edit
that object to include very complex stock dynamics if desired. Note that you can also create lists of these
stock objects if you want multiple species or growth-type-groups of the same species.
}
\examples{
A_short_lived_stock <- make_SL_stock("Default stock values for demo")
class(A_short_lived_stock)
}
\seealso{
\link{make_SL_fleet} for making a short-lived fleet object and \link{spec_SL_om} for specifying the entire operating model from fleet and stock objects.
}
\author{
T. Carruthers
}
