% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project_seasonal_recruitment.R
\name{Seasonality}
\alias{Seasonality}
\title{Takes an OpenMSE v1.0 object from RCM and imposes seasonality including estimated recruitment pattern in an OpenMSE v2.0 OM}
\usage{
Seasonality(OM, Seasons = 4, trim = 4, RecLev = 1, sy = 1, plot = F)
}
\arguments{
\item{OM}{An object of openMSE class 'OM' (OpenMSE v1.0)}

\item{Seasons}{Positive integer. The number of seasons within year. Defaults to 4, quarterly.}

\item{trim}{Positive integer. The number of recent historical years of estimated recruitment to ignore (due to poor characterization in the lower triangle).}

\item{RecLev}{Improper fraction. Overall mean recruitment strength relative to historical. Defaults to 1. Can be modified for purposes of OM specification.}

\item{sy}{Positive integer. The number of initial recruitment years (including age classes in initial year) to impose recruitment. Defaults to 0.}

\item{plot}{Boolean. Do you want to plot the estmated  / imposed recruitments?}
}
\description{
Takes an OpenMSE v1.0 object from RCM and imposes seasonality including estimated recruitment pattern in an OpenMSE v2.0 OM
}
\examples{
Seasonality(RCMfit@OM)
}
\author{
T. Carruthers
}
