% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slFleet.R
\name{slFleet}
\alias{slFleet}
\title{Make a Short-Lived Species Fleet Object}
\usage{
slFleet(
  Name = "A fleet",
  nYear = 10,
  pYear = 10,
  Seasons = 12,
  CurrentYear = 2026,
  nSim = 4,
  rec_age = 1,
  nAges = 24,
  Effort = NA,
  relE_loc = 0.35,
  relE_freq = 0.85,
  relE_min = 0.25,
  relE_max = 0.5,
  relE_CV = 0.25,
  maxF = 0.1,
  smu = 0.5,
  scv = 0.2,
  sel50 = 6,
  selSLP = 2,
  plot = F
)
}
\arguments{
\item{Name}{Character string. The name of the stock object. E.g., "North Pacific Neon Flying Squid"}

\item{nYear}{Positive Integer. The number of historical years. E.g., 10, (2015 - 2024)}

\item{pYear}{Positive Integer. The number of projection years. E.g., 10 (2025 - 2034)}

\item{Seasons}{Positive Integer. The number of sub-year time steps. E.g., 12 for a monthly model}

\item{CurrentYear}{Integer. The Calendar year of the last historical year. E.g. 2024.}

\item{nSim}{Positive Integer. The number of independent simulations. E.g., 4 for testing, 24 for preliminary results, 48 for representative results, 192+ for informing management.}

\item{rec_age}{Positive Integer. The age the fish is recruited (in seasonal time steps). E.g., given a monthly model (Seasons = 12) a value of 4 means that the stock recruitment function calculates recruiment into the time step 4 months later.}

\item{nAges}{Positive Integer. The number of seasonal time steps that population dynamics will be calculated for. E.g., 24 would be two years in a monthly model. Note that this only necessary if creatures live that long or dynamics are not suitably approximated with a plus group.}

\item{Effort}{A matrix of positive real numbers nSim x time steps (nYear x Seasons). Given a default q value of 1 this is the apical fishing mortality rate. Defaul is 'NA' and in this case effort pattern is simulated with a seasonal and temporal trend.}

\item{relE_loc}{Positive fraction. Starting location (year 1) of the annual sinewave pattern of effort.}

\item{relE_freq}{Positive real number. The frequency of the sine wave of annual effort. A value of 1 is one complete sinewave.}

\item{relE_min}{Positive real number. The minimum annual value of sin wave relative effort.}

\item{relE_max}{Positive real number. The maximum annual value of sin wave relative effort.}

\item{relE_CV}{Positive real number. The coefficient of variation for seasonal effort.}

\item{maxF}{Positive real number. The maximum apical (most selected age class) fishing mortality rate.}

\item{smu}{Positive fraction. The location of the mean (modal) seasonal effot within year. A value of 0.5 has a peak exactly mid year.}

\item{scv}{Positive real number. The CV controling the normal distribution of effort across seasons. A large value is 'flat' effort across seasons.}

\item{sel50}{Positive real number. The age (in seasons) that 50% of individuals are selected, in a logistic selectivity model. E.g., 6 in a monthly model would be 50% selected after 6 months.}

\item{selSLP}{Positive real number. The slope of the logistic selectivity model. E.g., 2 in a monthly model.}

\item{plot}{Boolean. Should plots be presented on the creation of the fleet object.}
}
\value{
An object of MSEtool class fleet
}
\description{
A function that builds the fleet component of an operating model for a short-lived species. Requires from
specifications for growth, stock-recruitment, natural mortality rate and spatial distribution. This
is a wrapper function and produces a fully customizable MSEtool object of class 'stock'. You can edit
that object to include very complex stock dynamics if desired.
}
\examples{
A_short_lived_fleet<- slFleet("Default fleet values for demo")
class(A_short_lived_fleet)
}
\seealso{
\link{slFleet} for making a short-lived stock object and \link{slOM} for specifying the entire operating model from fleet and stock objects.
}
\author{
T. Carruthers
}
