% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slOM.R
\name{slOM}
\alias{slOM}
\title{Construct Complete Short-Lived Species Operating Model}
\usage{
slOM(
  Name = "Short-lived simulation",
  Agency = "A fishery agency",
  Author = "A fishery analyst",
  Email = "a.person@email.com",
  Region = "A fishery management area",
  ComplexName = "Short-Lived Complex",
  Latitude = NA,
  Longitude = NA,
  Sponsor = "A generous funder",
  nSim = 4,
  nYear = 10,
  pYear = 10,
  Seasons = 12,
  CurrentYear = 2026,
  Interval = 6,
  Seed = 1,
  stock = NA,
  fleet = NA,
  obs = NA
)
}
\arguments{
\item{Name}{Character string. The name of the operating model. E.g., "NP NFS Ref OM #3: high M, high h, low rec"}

\item{Agency}{Character string. The relevant fishery agency "North Pacific Fishery Council (NPFC)"}

\item{Author}{Character string. Name of the author(s) of the operating model E.g., "A. Person"}

\item{Email}{Character string. Email address of the corresponding author}

\item{Region}{Character string. The location of the stock / management area.}

\item{ComplexName}{Character string. The name of the multi-stock complex (if more than 1 stock is being simulated).}

\item{Latitude}{Real number. Degrees N. Negative is southern hemisphere}

\item{Longitude}{Real number. Degrees E. Negative is western hemisphere}

\item{Sponsor}{Character string. The name of the person(s) or organization(s) that paid for the research.}

\item{nSim}{Positive Integer. The number of independent simulations. E.g., 4 for testing, 24 for preliminary results, 48 for representative results, 192+ for informing management.}

\item{nYear}{Positive Integer. The number of historical years. E.g., 10, (2015 - 2024)}

\item{pYear}{Positive Integer. The number of projection years. E.g., 10 (2025 - 2034)}

\item{Seasons}{Positive Integer. The number of sub-year time steps. E.g., 12 for a monthly model}

\item{CurrentYear}{Integer. The Calendar year of the last historical year. E.g. 2024.}

\item{Interval}{Positive integer. The management interval (in seasons) - how frequently is new advice provided. In a monthly model, a value of 6 would mean every 6 months after June and December.}

\item{Seed}{Real number. The seed for sampling of random variables.}

\item{stock}{A stock object or list of stock objects}

\item{fleet}{A fleet object or list of fleet objects}
}
\value{
An object of MSEtool class om
}
\description{
A function that combines stock and fleet object in a complete operating model for a short-lived species. Can
include lists of fleets and stocks for more complex operating models
}
\examples{
A_short_lived_om <- slOM("Default fleet values for demo")
class(A_short_lived_om)
}
\seealso{
\link{slStock} for making a short-lived stock object and \link{slFleet} for making a short-lived fleet object.
}
\author{
T. Carruthers
}
